app-automation/requirements.txt
Appium-Python-Client>=4.0,<5.0
PyYAML>=6.0
Pillow>=10.0
